Add Flow type checking to VerticalSlider.jsx Prior to this change, VerticalSlider.jsx did not have static type checking.
This change adds Flow type checking to VerticalSlider.jsx. One caveat is
that the value in VerticalSlider.jsx should probably be a number.
Currently, the envelopeSliderChange function in App.jsx uses ParseFloat
to make sure the value passed in is a number. However, it's probably
better practice to make sure it's a number in VerticalSlider.jsx first.
Nonetheless, all of the other type checking in VerticalSlider.jsx works,
so that can be adjusted at a later date.