Dynamically bind methods in Keyboard.jsx Prior to this change, there was a lot of boilerplate in the constructor for Keyboard.jsx. Also, the .filter method in 'const result =' returned an array that needed to be later parsed into its value.
This change has an array of method names and loops over them in the
constructor to dynamically create bound methods. This reduces the
boilerplate in the constructor code. Also, the .find method was swapped
in place of the .filter method to replace the need to later extract the
value from the array .filter would produce.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/find