Rename adsr to sliderName in slider props
Prior to this change, the name of the slider name prop passed
into slider components was "adsr." That made sense when it
was just a slider bank for the envelopes but now that it's
more reusable, it needs a better and more descriptive name.

This change changes the name from "adsr" to "sliderName" for better
readability.